Added the spring security JWT authorization

+++ src/main/java/com/aman/AirBnb/AirBnb/Service/HotelServiceImpl.java
@@ -5,7 +5,9 @@ import com.aman.AirBnb.AirBnb.Dto.HotelInfoDto;
 import com.aman.AirBnb.AirBnb.Dto.RoomDTO;
 import com.aman.AirBnb.AirBnb.Entities.HotelEntity;
 import com.aman.AirBnb.AirBnb.Entities.RoomEntity;
+import com.aman.AirBnb.AirBnb.Entities.UserEntity;
 import com.aman.AirBnb.AirBnb.Exceptions.ResourceNotFoundException;
+import com.aman.AirBnb.AirBnb.Exceptions.UnAuthorisedException;
 import com.aman.AirBnb.AirBnb.Repositories.HotelRepository;
 import com.aman.AirBnb.AirBnb.Repositories.RoomRepository;
 import com.aman.AirBnb.AirBnb.Service.Interfaces.HotelService;
@@ -13,6 +15,7 @@ import com.aman.AirBnb.AirBnb.Service.Interfaces.InventoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,6 +36,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         HotelEntity hotel = modelMapper.map(hotelDto, HotelEntity.class);
         hotel.setActive(false);
+
+        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotel.getId());
         return modelMapper.map(hotel, HotelDTO.class);
@@ -44,6 +51,13 @@ public class HotelServiceImpl implements HotelService {
         HotelEntity hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         return modelMapper.map(hotel, HotelDTO.class);
     }
 
@@ -53,6 +67,12 @@ public class HotelServiceImpl implements HotelService {
         HotelEntity hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -65,6 +85,11 @@ public class HotelServiceImpl implements HotelService {
         HotelEntity hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
 
         for(RoomEntity room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
@@ -81,14 +106,19 @@ public class HotelServiceImpl implements HotelService {
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
-        hotel.setActive(true);
+        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
 
+        hotel.setActive(true);
         // assuming only do it once
         for(RoomEntity room: hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
         }
     }
 
+    //public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         HotelEntity hotel = hotelRepository
